shazam/requirements.txt
amqp==2.5.0
anyjson==0.3.3
backports.functools-lru-cache==1.5
billiard==3.6.0.0
celery==4.3.0
certifi==2019.6.16
chardet==3.0.4
cycler==0.10.0
Django==1.11
django-celery==3.2.2
django-markdown==0.8.4
djangorestframework==3.4.6
idna==2.8
kiwisolver==1.1.0
kombu==4.6.3
Markdown==3.1.1
matplotlib==2.2.4
MySQL-python==1.2.5
numpy==1.16.4
PyAudio==0.2.11
pydub==0.23.1
pyfcm==1.4.7
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2019.1
requests==2.22.0
scipy==1.2.2
six==1.12.0
subprocess32==3.5.4
urllib3==1.25.3
vine==1.3.0
